Move the logic from `EvaluatorCommand` to a new module `evaluator`
This makes it easier to test the code.
