+++ evaluator/build.gradle
@@ -1,0 +1,12 @@
+// Apply core plugins.
+apply plugin: 'java-library'
+
+dependencies {
+    api project(':model')
+
+    implementation project(':utils')
+
+    implementation 'org.jetbrains.kotlin:kotlin-compiler-embeddable'
+    implementation 'org.jetbrains.kotlin:kotlin-script-runtime'
+    implementation 'org.jetbrains.kotlin:kotlin-script-util'
+}
