+++ evaluator/src/main/resources/rules/no_gpl_declared.kts
