+++ settings.gradle
@@ -25,6 +25,7 @@ rootProject.name = 'oss-review-toolkit'
 include 'analyzer'
 include 'cli'
 include 'downloader'
+include 'evaluator'
 include 'model'
 include 'reporter'
 include 'reporter-web-app'
